+++ receiver/swok8sdiscovery/factory.go
@@ -34,16 +34,8 @@ func NewFactory() receiver.Factory {
 	)
 }
 
-func createDefaultConfig() component.Config {
-	return &Config{
-		APIConfig: k8sconfig.APIConfig{
-			AuthType: k8sconfig.AuthTypeServiceAccount,
-		},
-	}
-}
-
 // Build-in configuration
-func createImplicitConfig() component.Config {
+func createDefaultConfig() component.Config {
 	return &Config{
 		APIConfig: k8sconfig.APIConfig{
 			AuthType: k8sconfig.AuthTypeServiceAccount,
@@ -52,27 +44,27 @@ func createImplicitConfig() component.Config {
 			ImageRules: []*ImageRule{
 				{
 					DatabaseType: "redis",
-					Patterns:     []string{".*/redis:.*"},
+					Patterns:     []string{`(^|/)redis:.*`},
 					DefaultPort:  6379,
 				},
 				{
 					DatabaseType: "mysql",
-					Patterns:     []string{".*/mysql:.*", ".*/mariadb:.*", ".*/mysql-server:.*"},
+					Patterns:     []string{`(^|/)mysql:.*`, `(^|/)mariadb:.*`, `(^|/)mysql-server:.*`},
 					DefaultPort:  3306,
 				},
 				{
 					DatabaseType: "sqlserver",
-					Patterns:     []string{".*/mssql:.*", ".*/mssql/server:.*"},
+					Patterns:     []string{`(^|/)mssql:.*`, `(^|/)mssql/server:.*`},
 					DefaultPort:  1433,
 				},
 				{
 					DatabaseType: "postgresql",
-					Patterns:     []string{".*/postgres:.*", ".*/postgresql:.*"},
+					Patterns:     []string{`(^|/)postgres:.*`, `(^|/)postgresql:.*`},
 					DefaultPort:  5432,
 				},
 				{
 					DatabaseType: "mongodb",
-					Patterns:     []string{".*/mongo:.*"},
+					Patterns:     []string{`(^|/)mongo:.*`},
 					DefaultPort:  27017,
 				},
 			},