+++ connector/solarwindsentityconnector/internal/event_detector_test.go
@@ -48,17 +48,17 @@ func TestDetect_EntityAndRelationshipEvents(t *testing.T) {
 		IDs:        []string{"id"},
 		Attributes: []string{"attr"},
 	}
-	entityEvent := config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{
+	entityEvent := config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{
 		Definition:   &config.EntityEvent{Entity: entity.Entity, Event: config.Event{Action: EventUpdateAction}},
 		ConditionSeq: seq,
 	}
-	relationshipEvent := config.ParsedEvent[config.RelationshipEvent, ottllog.TransformContext]{
+	relationshipEvent := config.ParsedEvent[config.RelationshipEvent, *ottllog.TransformContext]{
 		Definition:   &config.RelationshipEvent{Type: relationshipType, Source: entity.Entity, Destination: entity.Entity, Event: config.Event{Action: EventUpdateAction}},
 		ConditionSeq: seq,
 	}
-	eventsGroup := config.EventsGroup[ottllog.TransformContext]{
-		Entities:      []config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{entityEvent},
-		Relationships: []config.ParsedEvent[config.RelationshipEvent, ottllog.TransformContext]{relationshipEvent},
+	eventsGroup := config.EventsGroup[*ottllog.TransformContext]{
+		Entities:      []config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{entityEvent},
+		Relationships: []config.ParsedEvent[config.RelationshipEvent, *ottllog.TransformContext]{relationshipEvent},
 	}
 
 	// Prepare resource attributes
@@ -82,9 +82,7 @@ func TestDetect_EntityAndRelationshipEvents(t *testing.T) {
 	rLogs := logs.ResourceLogs().AppendEmpty()
 	scopeLogs := rLogs.ScopeLogs().AppendEmpty()
 	logRecord := scopeLogs.LogRecords().AppendEmpty()
-	resource := rLogs.Resource()
-	scope := scopeLogs.Scope()
-	tc := ottllog.NewTransformContext(logRecord, scope, resource, scopeLogs, rLogs)
+	tc := ottllog.NewTransformContextPtr(rLogs, scopeLogs, logRecord)
 
 	attributeMapper := NewAttributeMapper(map[string]config.Entity{entity.Entity: entity})
 
@@ -117,12 +115,12 @@ func TestDetect_NoEvents(t *testing.T) {
 		IDs:        []string{"id"},
 		Attributes: []string{"attr"},
 	}
-	entityEvent := config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{
+	entityEvent := config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{
 		Definition:   &config.EntityEvent{Entity: entity.Entity},
 		ConditionSeq: seq,
 	}
-	eventsGroup := config.EventsGroup[ottllog.TransformContext]{
-		Entities:      []config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{entityEvent},
+	eventsGroup := config.EventsGroup[*ottllog.TransformContext]{
+		Entities:      []config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{entityEvent},
 		Relationships: nil,
 	}
 
@@ -137,9 +135,7 @@ func TestDetect_NoEvents(t *testing.T) {
 	rLogs := logs.ResourceLogs().AppendEmpty()
 	scopeLogs := rLogs.ScopeLogs().AppendEmpty()
 	logRecord := scopeLogs.LogRecords().AppendEmpty()
-	resource := rLogs.Resource()
-	scope := scopeLogs.Scope()
-	tc := ottllog.NewTransformContext(logRecord, scope, resource, scopeLogs, rLogs)
+	tc := ottllog.NewTransformContextPtr(rLogs, scopeLogs, logRecord)
 
 	attributeMapper := NewAttributeMapper(map[string]config.Entity{entity.Entity: entity})
 
@@ -167,25 +163,23 @@ func TestProcessEvents_ConditionTrue_EventsCreated(t *testing.T) {
 	require.NoError(t, err)
 	seq := ottl.NewConditionSequence(stmts, settings)
 
-	entityEvent := config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{
+	entityEvent := config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{
 		Definition:   &config.EntityEvent{Entity: "test-entity"},
 		ConditionSeq: seq,
 	}
-	relationshipEvent := config.ParsedEvent[config.RelationshipEvent, ottllog.TransformContext]{
+	relationshipEvent := config.ParsedEvent[config.RelationshipEvent, *ottllog.TransformContext]{
 		Definition:   &config.RelationshipEvent{Type: "test-rel"},
 		ConditionSeq: seq,
 	}
-	eventsGroup := config.EventsGroup[ottllog.TransformContext]{
-		Entities:      []config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{entityEvent},
-		Relationships: []config.ParsedEvent[config.RelationshipEvent, ottllog.TransformContext]{relationshipEvent},
+	eventsGroup := config.EventsGroup[*ottllog.TransformContext]{
+		Entities:      []config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{entityEvent},
+		Relationships: []config.ParsedEvent[config.RelationshipEvent, *ottllog.TransformContext]{relationshipEvent},
 	}
 
 	rLogs := logs.ResourceLogs().AppendEmpty()
 	scopeLogs := rLogs.ScopeLogs().AppendEmpty()
 	logRecord := scopeLogs.LogRecords().AppendEmpty()
-	resource := rLogs.Resource()
-	scope := scopeLogs.Scope()
-	tc := ottllog.NewTransformContext(logRecord, scope, resource, scopeLogs, rLogs)
+	tc := ottllog.NewTransformContextPtr(rLogs, scopeLogs, logRecord)
 	am := NewAttributeMapper(map[string]config.Entity{})
 	ev := NewEventDetector(am, eventsGroup, zap.NewNop())
 
@@ -209,25 +203,23 @@ func TestProcessEvents_ConditionFalse_EventsNotCreated(t *testing.T) {
 	require.NoError(t, err)
 	seq := ottl.NewConditionSequence(stmts, settings)
 
-	entityEvent := config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{
+	entityEvent := config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{
 		Definition:   &config.EntityEvent{Entity: "test-entity"},
 		ConditionSeq: seq,
 	}
-	relationshipEvent := config.ParsedEvent[config.RelationshipEvent, ottllog.TransformContext]{
+	relationshipEvent := config.ParsedEvent[config.RelationshipEvent, *ottllog.TransformContext]{
 		Definition:   &config.RelationshipEvent{Type: "test-rel"},
 		ConditionSeq: seq,
 	}
-	eventsGroup := config.EventsGroup[ottllog.TransformContext]{
-		Entities:      []config.ParsedEvent[config.EntityEvent, ottllog.TransformContext]{entityEvent},
-		Relationships: []config.ParsedEvent[config.RelationshipEvent, ottllog.TransformContext]{relationshipEvent},
+	eventsGroup := config.EventsGroup[*ottllog.TransformContext]{
+		Entities:      []config.ParsedEvent[config.EntityEvent, *ottllog.TransformContext]{entityEvent},
+		Relationships: []config.ParsedEvent[config.RelationshipEvent, *ottllog.TransformContext]{relationshipEvent},
 	}
 
 	rLogs := logs.ResourceLogs().AppendEmpty()
 	scopeLogs := rLogs.ScopeLogs().AppendEmpty()
 	logRecord := scopeLogs.LogRecords().AppendEmpty()
-	resource := rLogs.Resource()
-	scope := scopeLogs.Scope()
-	tc := ottllog.NewTransformContext(logRecord, scope, resource, scopeLogs, rLogs)
+	tc := ottllog.NewTransformContextPtr(rLogs, scopeLogs, logRecord)
 	am := NewAttributeMapper(map[string]config.Entity{})
 	ev := NewEventDetector(am, eventsGroup, zap.NewNop())
 
@@ -274,7 +266,7 @@ func TestGetRelationships_WithDifferentTypes(t *testing.T) {
 
 	eventDetector := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -319,7 +311,7 @@ func TestGetRelationships_WithSameType(t *testing.T) {
 
 	eventDetector := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -353,7 +345,7 @@ func TestCollectEvents_WithEntity_AttributesPresent(t *testing.T) {
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -394,7 +386,7 @@ func TestCollectEvents_WithEntity_IDAttributesMissing(t *testing.T) {
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -420,7 +412,7 @@ func TestCollectEvents_WithEntity_SomeAttributesMissing(t *testing.T) {
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -460,7 +452,7 @@ func TestCollectEvents_WithRelationship_AttributesPresent(t *testing.T) {
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -506,7 +498,7 @@ func TestCollectEvents_WithRelationship_SameType_AttributesPresent(t *testing.T)
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -552,7 +544,7 @@ func TestCollectEvents_WithRelationship_IDAttributesMissing(t *testing.T) {
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 
@@ -584,7 +576,7 @@ func TestCollectEvents_WithRelationship_SameType_IDAttributesMissing(t *testing.
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 	events, err := eventBuilder.collectEvents(
@@ -624,7 +616,7 @@ func TestCollectEvents_WithRelationship_RelationshipAttributesPresent(t *testing
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 	events, err := eventBuilder.collectEvents(
@@ -673,7 +665,7 @@ func TestCollectEvents_WithRelationship_SameType_RelationshipAttributesPresent(t
 
 	eventBuilder := NewEventDetector(
 		attributeMapper,
-		config.EventsGroup[ottllog.TransformContext]{},
+		config.EventsGroup[*ottllog.TransformContext]{},
 		zap.NewNop(),
 	)
 	events, err := eventBuilder.collectEvents(